requirements.txt
alembic==1.9.2
annotated-types==0.7.0
anyio==4.9.0
boto3==1.40.1
botocore==1.40.1
cachetools==5.5.2
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.1.3
distro==1.9.0
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
exceptiongroup==1.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
Flask-Migrate==4.0.2
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.1.1
google-auth==2.40.3
greenlet==3.0.1
gunicorn==20.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib-metadata==6.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.10.0
jmespath==1.0.1
lxml==6.0.0
Mako==1.2.4
MarkupSafe==2.1.2
openai==1.98.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.3
python-dateutil==2.8.2
python-docx==1.2.0
python-dotenv==0.21.0
python-editor==1.0.4
python-jose==3.5.0
requests==2.32.4
rsa==4.9.1
s3transfer==0.13.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.46
stripe==12.4.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==1.26.20
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.17.0
